feat(vad-trt): move VadMode to hpp
