+++ AV-Solutions/vad-trt/app/lib/vad_model.hpp
@@ -52,6 +52,31 @@ public:
   void error(const std::string& message) { log(LogLevel::ERROR, message); }
 };
 
+// シンプルなロガー実装（内部使用）
+class SimpleVadLogger : public VadLogger {
+public:
+    void log(LogLevel level, const std::string& message) override {
+        // 何もしない（サイレント）、または必要に応じて実装
+        // RosVadLoggerがデフォルトとして推奨される
+    }
+};
+
+// Loggerクラス（VadModel内で使用）
+class Logger : public nvinfer1::ILogger {
+private:
+    std::shared_ptr<VadLogger> custom_logger_;
+
+public:
+    Logger(std::shared_ptr<VadLogger> logger) : custom_logger_(logger) {}
+    
+    void log(nvinfer1::ILogger::Severity severity, const char* msg) noexcept override {
+        // Only print error messages
+        if (severity == nvinfer1::ILogger::Severity::kERROR && custom_logger_) {
+            custom_logger_->error(std::string(msg));
+        }
+    }
+};
+
 // VAD推論の入力データ構造
 struct VadInputData
 {
@@ -131,19 +156,97 @@ class VadModel
 {
 public:
   // コンストラクタ（サイレントロガーを使用）
-  VadModel(const VadConfig& config);
+  VadModel(const VadConfig& config)
+    : initialized_(false), stream_(nullptr), is_first_frame_(true), config_(config)
+  {
+    // シンプルなロガーをデフォルトとして使用（サイレント）
+    // 実際の使用ではRosVadLoggerを推奨
+    logger_ = std::make_shared<SimpleVadLogger>();
+    
+    // 初期化を実行
+    runtime_ = create_runtime();
+    
+    if (!load_plugin(config.plugins_path)) {
+      logger_->error("Failed to load plugin");
+      return;
+    }
+    
+    cudaStreamCreate(&stream_);
+    
+    nets_ = init_engines(config.nets_config);
+    
+    logger_->info("warm_up=" + std::to_string(config.warm_up_num));
+    warm_up(config.warm_up_num);
+    
+    initialized_ = true;
+  }
 
   // ロガー付きコンストラクタ（推奨）
-  VadModel(const VadConfig& config, std::shared_ptr<VadLogger> logger);
+  VadModel(const VadConfig& config, std::shared_ptr<VadLogger> logger)
+    : initialized_(false), stream_(nullptr), is_first_frame_(true), config_(config), logger_(logger)
+  {
+    // 初期化を実行
+    runtime_ = create_runtime();
+    
+    if (!load_plugin(config.plugins_path)) {
+      logger_->error("Failed to load plugin");
+      return;
+    }
+    
+    cudaStreamCreate(&stream_);
+    
+    nets_ = init_engines(config.nets_config);
+    
+    logger_->info("warm_up=" + std::to_string(config.warm_up_num));
+    warm_up(config.warm_up_num);
+    
+    initialized_ = true;
+  }
 
   // デストラクタ
-  ~VadModel();
-
-  // モデルの初期化
-  [[nodiscard]] bool initialize(const std::string & model_path);
+  ~VadModel()
+  {
+    if (stream_) {
+      cudaStreamDestroy(stream_);
+      stream_ = nullptr;
+    }
+    
+    // netsのクリーンアップ
+    nets_.clear();
+    
+    initialized_ = false;
+  }
 
   // メイン推論API
-  [[nodiscard]] std::optional<VadOutputData> infer(const VadInputData & input);
+  [[nodiscard]] std::optional<VadOutputData> infer(const VadInputData & vad_input) {
+    // 最初のフレームかどうかでheadの名前を変更
+    std::string head_name;
+    if (is_first_frame_) {
+      head_name = "head_no_prev";
+    } else {
+      head_name = "head";
+    }
+
+    // bindingsにload
+    load_inputs(vad_input, head_name);
+
+    // backboneとheadをenqueue
+    enqueue(head_name);
+
+    // prev_bevを保存
+    saved_prev_bev_ = save_prev_bev(head_name);
+    // VadOutputDataに出力を変換
+    VadOutputData output = postprocess(head_name, vad_input.command_);
+
+    // 最初のフレームなら以下の処理を行う
+    if (is_first_frame_) {
+      release_network("head_no_prev");
+      load_head();
+      is_first_frame_ = false;
+    }
+    
+    return output;
+  }
 
   // メンバ変数
   std::unique_ptr<nvinfer1::IRuntime, std::function<void(nvinfer1::IRuntime*)>> runtime_;
@@ -163,19 +266,152 @@ public:
 
 private:
   // メンバ関数
-  std::unique_ptr<nvinfer1::IRuntime, std::function<void(nvinfer1::IRuntime*)>> create_runtime();
-  bool load_plugin(const std::string& plugin_dir);
+  std::unique_ptr<nvinfer1::IRuntime, std::function<void(nvinfer1::IRuntime*)>> create_runtime() {
+    static std::unique_ptr<Logger> logger_instance = std::make_unique<Logger>(logger_);
+    auto runtime_deleter = [](nvinfer1::IRuntime *runtime) {};
+    std::unique_ptr<nvinfer1::IRuntime, decltype(runtime_deleter)> runtime{
+        nvinfer1::createInferRuntime(*logger_instance), runtime_deleter};
+    return runtime;
+  }
+
+  bool load_plugin(const std::string& plugin_dir) {
+    void* h_ = dlopen(plugin_dir.c_str(), RTLD_NOW);
+    logger_->info("loading plugin from: " + plugin_dir);
+    if (!h_) {
+      const char* error = dlerror();
+      logger_->error("Failed to load library: " + std::string(error ? error : "unknown error"));
+      return false;
+    }
+    return true;
+  }
+
   std::unordered_map<std::string, std::shared_ptr<nv::Net>> init_engines(
-    const std::vector<NetConfig>& nets_config);
-  void warm_up(int32_t warm_up_num);
+      const std::vector<NetConfig>& nets_config) {
+    
+    std::unordered_map<std::string, std::shared_ptr<nv::Net>> nets;
+    
+    for (const auto& engine : nets_config) {
+      if (engine.name == "head") {
+        continue;  // headは後で初期化
+      }
+      
+      std::string engine_name = engine.name;
+      std::string engine_file_path = engine.engine_file;
+      logger_->info("-> engine: " + engine_name);
+      
+      std::unordered_map<std::string, std::shared_ptr<nv::Tensor>> external_bindings;
+      // reuse memory
+      for (const auto& input_pair : engine.inputs) {
+        const std::string& k = input_pair.first;
+        const auto& ext_map = input_pair.second;      
+        std::string ext_net = ext_map.at("net");
+        std::string ext_name = ext_map.at("name");
+        logger_->info(k + " <- " + ext_net + "[" + ext_name + "]");
+        external_bindings[k] = nets[ext_net]->bindings[ext_name];
+      }
+
+      nets[engine_name] = std::make_shared<nv::Net>(engine_file_path, runtime_.get(), external_bindings);
+
+      if (engine.use_graph) {
+        nets[engine_name]->EnableCudaGraph(stream_);
+      }
+    }
+    
+    return nets;
+  }
+
+  void warm_up(int32_t warm_up_num) {
+    for(int32_t iw=0; iw < warm_up_num; iw++) {
+      nets_["backbone"]->Enqueue(stream_);
+      nets_["head_no_prev"]->Enqueue(stream_);
+      cudaStreamSynchronize(stream_);
+    }
+  }
   
   // infer関数で使用するヘルパー関数
-  void load_inputs(const VadInputData& vad_input, const std::string& head_name);
-  void enqueue(const std::string& head_name);
-  std::shared_ptr<nv::Tensor> save_prev_bev(const std::string& head_name);
-  void release_network(const std::string& network_name);
-  void load_head();
-  VadOutputData postprocess(const std::string& head_name, int32_t cmd);
+  void load_inputs(const VadInputData& vad_input, const std::string& head_name) {
+    nets_["backbone"]->bindings["img"]->load(vad_input.camera_images_, stream_);
+    nets_[head_name]->bindings["img_metas.0[shift]"]->load(vad_input.shift_, stream_);
+    nets_[head_name]->bindings["img_metas.0[lidar2img]"]->load(vad_input.lidar2img_, stream_);
+    nets_[head_name]->bindings["img_metas.0[can_bus]"]->load(vad_input.can_bus_, stream_);
+
+    if (head_name == "head") {
+      nets_["head"]->bindings["prev_bev"] = saved_prev_bev_;
+    }
+  }
+
+  void enqueue(const std::string& head_name) {
+    nets_["backbone"]->Enqueue(stream_);
+    nets_[head_name]->Enqueue(stream_);
+  }
+
+  std::shared_ptr<nv::Tensor> save_prev_bev(const std::string& head_name) {
+    auto bev_embed = nets_[head_name]->bindings["out.bev_embed"];
+    auto prev_bev = std::make_shared<nv::Tensor>("prev_bev", bev_embed->dim, bev_embed->dtype);
+    cudaMemcpyAsync(prev_bev->ptr, bev_embed->ptr, bev_embed->nbytes(), 
+                    cudaMemcpyDeviceToDevice, stream_);
+    return prev_bev;
+  }
+
+  void release_network(const std::string& network_name) {
+    if (nets_.find(network_name) != nets_.end()) {
+      // まずbindingsをクリア
+      nets_[network_name]->bindings.clear();
+      cudaDeviceSynchronize();
+      
+      // 次にNetオブジェクトを解放
+      nets_[network_name].reset();
+      nets_.erase(network_name);
+      cudaDeviceSynchronize();
+    }
+  }
+
+  void load_head() {
+    auto head_engine = std::find_if(config_.nets_config.begin(), config_.nets_config.end(),
+        [](const NetConfig& engine) { return engine.name == "head"; });
+    
+    if (head_engine == config_.nets_config.end()) {
+      logger_->error("Head engine configuration not found");
+      return;
+    }
+    
+    std::string engine_file_path = head_engine->engine_file;
+    logger_->info("-> loading head engine: " + engine_file_path);
+    
+    std::unordered_map<std::string, std::shared_ptr<nv::Tensor>> external_bindings;
+    for (const auto& input_pair : head_engine->inputs) {
+      const std::string& k = input_pair.first;
+      const auto& ext_map = input_pair.second;      
+      std::string ext_net = ext_map.at("net");
+      std::string ext_name = ext_map.at("name");
+      logger_->info(k + " <- " + ext_net + "[" + ext_name + "]");
+      external_bindings[k] = nets_[ext_net]->bindings[ext_name];
+    }
+
+    nets_["head"] = std::make_shared<nv::Net>(engine_file_path, runtime_.get(), external_bindings);
+
+    if (head_engine->use_graph) {
+      nets_["head"]->EnableCudaGraph(stream_);
+    }
+  }
+
+  VadOutputData postprocess(const std::string& head_name, int32_t cmd) {
+    std::vector<float> ego_fut_preds = nets_[head_name]->bindings["out.ego_fut_preds"]->cpu<float>();
+    
+    // Extract planning for the given command
+    std::vector<float> planning(
+        ego_fut_preds.begin() + cmd * 12,
+        ego_fut_preds.begin() + (cmd + 1) * 12
+    );
+    
+    // cumsum to build trajectory in 3d space
+    for (int32_t i = 1; i < 6; i++) {
+      planning[i * 2] += planning[(i-1) * 2];
+      planning[i * 2 + 1] += planning[(i-1) * 2 + 1];
+    }
+    
+    return VadOutputData{planning};
+  }
 };
 
 }  // namespace autoware::tensorrt_vad
